fix: Remove emotes v2, smart items and worlds for ens owners FFs
