+++ src/components/InspectorPage/InspectorPage.container.ts
@@ -6,7 +6,6 @@ import { getCurrentScene } from 'modules/scene/selectors'
 import { getCurrentProject } from 'modules/project/selectors'
 import { connectInspector, openInspector } from 'modules/inspector/actions'
 import { isReloading } from 'modules/inspector/selectors'
-import { getIsSmartItemsEnabled } from 'modules/features/selectors'
 import { MapStateProps, MapDispatch, MapDispatchProps } from './InspectorPage.types'
 import EditorPage from './InspectorPage'
 
@@ -16,8 +15,7 @@ const mapState = (state: RootState): MapStateProps => {
     isLoggedIn: isLoggedIn(state),
     scene: getCurrentScene(state),
     project: getCurrentProject(state),
-    isReloading: isReloading(state),
-    isSmartItemsEnabled: getIsSmartItemsEnabled(state)
+    isReloading: isReloading(state)
   }
 }
