+++ src/components/InspectorPage/InspectorPage.tsx
@@ -30,7 +30,7 @@ export default class InspectorPage extends React.PureComponent<Props, State> {
   }
 
   render() {
-    const { scene, project, isLoggedIn, isReloading, isSmartItemsEnabled, address = '' } = this.props
+    const { scene, project, isLoggedIn, isReloading, address = '' } = this.props
 
     if (!isLoggedIn) {
       return <SignInRequired />
@@ -65,10 +65,6 @@ export default class InspectorPage extends React.PureComponent<Props, State> {
 
     if (project?.id) {
       queryParams = queryParams.concat(`&projectId=${project.id}`)
-    }
-
-    if (!isSmartItemsEnabled) {
-      queryParams = queryParams.concat('&disableSmartItems')
     }
 
     return (
