+++ src/components/Modals/DeployModal/DeployModal.container.ts
@@ -4,7 +4,6 @@ import { getCurrentProject } from 'modules/project/selectors'
 import { getData as getDeployments } from 'modules/deployment/selectors'
 import { getActivePoolGroup } from 'modules/poolGroup/selectors'
 import { getCurrentScene } from 'modules/scene/selectors'
-import { getIsWorldsForEnsOwnersEnabled } from 'modules/features/selectors'
 import { MapStateProps, OwnProps } from './DeployModal.types'
 import DeployModal from './DeployModal'
 
@@ -12,8 +11,7 @@ const mapState = (state: RootState, ownProps: OwnProps): MapStateProps => ({
   deployment: getDeployments(state)[ownProps.metadata.projectId],
   currentPoolGroup: getActivePoolGroup(state),
   project: getCurrentProject(state),
-  scene: getCurrentScene(state),
-  isWorldsForEnsOwnersEnabled: getIsWorldsForEnsOwnersEnabled(state)
+  scene: getCurrentScene(state)
 })
 
 export default connect(mapState)(DeployModal)
