+++ src/components/Modals/DeployModal/DeployModal.tsx
@@ -6,7 +6,6 @@ import Icon from 'components/Icon'
 import DeployToLand from './DeployToLand'
 import DeployToPool from './DeployToPool'
 import DeployToWorld from './DeployToWorld'
-import DeployToWorldWorldsForEnsOwners from './DeployToWorld_WorldsForEnsOwnersFeature'
 import ClearDeployment from './ClearDeployment'
 import { Props, State, DeployModalView } from './DeployModal.types'
 import './DeployModal.css'
@@ -133,7 +132,7 @@ export default class DeployModal extends React.PureComponent<Props, State> {
 
   render() {
     const { view, deploymentId, claimedName } = this.state
-    const { name, currentPoolGroup, scene, isWorldsForEnsOwnersEnabled } = this.props
+    const { name, currentPoolGroup, scene } = this.props
 
     if (view === DeployModalView.CLEAR_DEPLOYMENT && deploymentId) {
       return <ClearDeployment deploymentId={deploymentId} name={name} onClose={this.handleClose} />
@@ -157,11 +156,7 @@ export default class DeployModal extends React.PureComponent<Props, State> {
     }
 
     if (view === DeployModalView.DEPLOY_TO_WORLD) {
-      return isWorldsForEnsOwnersEnabled ? (
-        <DeployToWorldWorldsForEnsOwners claimedName={claimedName} name={name} onClose={this.handleClose} onBack={this.handleBack} />
-      ) : (
-        <DeployToWorld claimedName={claimedName} name={name} onClose={this.handleClose} onBack={this.handleBack} />
-      )
+      return <DeployToWorld claimedName={claimedName} name={name} onClose={this.handleClose} onBack={this.handleBack} />
     }
 
     return this.renderChoiceForm()
