--- src/components/Modals/DeployModal/DeployToWorld_WorldsForEnsOwnersFeature/index.ts
@@ -1,3 +1,0 @@
-import DeployToWorld from './DeployToWorld.container'
-
-export default DeployToWorld
