+++ src/components/ScenesPage/ScenesPage.types.ts
@@ -28,7 +28,7 @@ export type Props = DefaultProps & {
 
 export type MapStateProps = Pick<
   Props,
-  'projects' | 'isFetching' | 'isLoggingIn' | 'page' | 'sortBy' | 'totalPages' | 'didCreate' | 'poolList' | 'isWorldsForENSOwnersEnabled'
+  'projects' | 'isFetching' | 'isLoggingIn' | 'page' | 'sortBy' | 'totalPages' | 'didCreate' | 'poolList'
 >
 export type MapDispatchProps = Pick<Props, 'onOpenModal' | 'onPageChange' | 'onLoadFromScenePool'>
 export type MapDispatch = Dispatch<CallHistoryMethodAction | OpenModalAction | LoadPoolsRequestAction>
