+++ src/modules/features/selectors.spec.ts
@@ -6,7 +6,6 @@ import {
   getIsCreateSceneOnlySDK7Enabled,
   getIsMaintenanceEnabled,
   getIsSDK7TemplatesEnabled,
-  getIsWorldsForEnsOwnersEnabled,
   getIsNavbarV2Enabled,
   getIsEnsAddressEnabled
 } from './selectors'
@@ -62,7 +61,6 @@ describe('when getting if maintainance is enabled', () => {
 })
 
 const ffSelectors = [
-  { selector: getIsWorldsForEnsOwnersEnabled, app: ApplicationName.BUILDER, feature: FeatureName.WORLDS_FOR_ENS_OWNERS },
   { selector: getIsSDK7TemplatesEnabled, app: ApplicationName.BUILDER, feature: FeatureName.SDK7_TEMPLATES },
   { selector: getIsCreateSceneOnlySDK7Enabled, app: ApplicationName.BUILDER, feature: FeatureName.CREATE_SCENE_ONLY_SDK7 },
   { selector: getIsAuthDappEnabled, app: ApplicationName.DAPPS, feature: FeatureName.AUTH_DAPP },
